+++ JLiteralLong.java
@@ -23,7 +23,7 @@ class JLiteralLong extends JExpression {
      */
 
     public JExpression analyze(Context context) {
-        type = Type.INT;
+        type = Type.LONG;
         return this;
     }
